% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SidraFacil.R
\name{sidra_facil}
\alias{sidra_facil}
\title{SIDRA FACIL - Pacote Framework para baixa automatizada de dados do SIDRA IBGE
A proposta deste pacote framework - proporcionar uma ferramenta FACIL e descomplicada para baixa de dados do SIDRA IBGE a partir de simples selecao pelo usuario
Solicita um titulo e um texto ao usuario e salva em um arquivo TXT.}
\usage{
sidra_facil()
}
\value{
Caminho completo do arquivo criado.
}
\description{
SIDRA FACIL - Pacote Framework para baixa automatizada de dados do SIDRA IBGE
A proposta deste pacote framework - proporcionar uma ferramenta FACIL e descomplicada para baixa de dados do SIDRA IBGE a partir de simples selecao pelo usuario
Solicita um titulo e um texto ao usuario e salva em um arquivo TXT.
}
\examples{
\dontrun{
dados <- sidra_facil(
  tabela = 1419,
  variavel = 93
)
}

}
